.history/backend/requirements_20251213112254.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
pydantic-settings==2.5.2
feedparser==6.0.11
httpx==0.27.2
python-dateutil==2.9.0.post0
APScheduler==3.10.4
beautifulsoup4==4.12.2
